feat: move filter functions and SingleSubNoEoseCacheEoseManager to commons
Move 15 files to commons module for KMP iOS compatibility:

Eose Managers (1):
- SingleSubNoEoseCacheEoseManager → commons.relayClient.eoseManagers

Filter Functions (14):
- FilterPostsByScopes (CommentKinds) → commons.relayClient.filters
- FilterCommunityPosts → commons.relayClient.filters
- FilterPostsByHashtags → commons.relayClient.filters
- FilterPostsByGeohash → commons.relayClient.filters
- FilterPostsByRelay → commons.relayClient.filters
- FilterUserProfileLastSeen → commons.relayClient.filters
- FilterUserProfileLists → commons.relayClient.filters
- FilterUserProfileMetadata → commons.relayClient.filters
- FilterMessagesToEphemeralChat → commons.relayClient.filters
- FilterMessagesToLiveStream → commons.relayClient.filters
- FilterMessagesToPublicChat → commons.relayClient.filters
- FilterMyMessagesToEphemeralChat → commons.relayClient.filters
- FilterMyMessagesToLiveActivities → commons.relayClient.filters
- FilterMyMessagesToPublicChat → commons.relayClient.filters

All original files replaced with re-export wrappers for backward
compatibility. Both :commons:compileKotlinJvm and
:amethyst:compilePlayDebugKotlin pass.
