+++ amethyst/src/main/java/com/vitorpamplona/amethyst/service/relayClient/eoseManagers/SingleSubNoEoseCacheEoseManager.kt
@@ -20,59 +20,4 @@
  */
 package com.vitorpamplona.amethyst.service.relayClient.eoseManagers
 
-import com.vitorpamplona.amethyst.commons.relayClient.eoseManagers.BaseEoseManager
-import com.vitorpamplona.quartz.nip01Core.core.Event
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.groupByRelay
-import com.vitorpamplona.quartz.nip01Core.relay.client.reqs.SubscriptionListener
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-
-/**
- * This query type creates only ONE relay subscription and does not
- * store any EOSE cache. The EOSE is irrelevant for these types of
- * filters
- */
-abstract class SingleSubNoEoseCacheEoseManager<T>(
-    client: INostrClient,
-    allKeys: () -> Set<T>,
-    val invalidateAfterEose: Boolean = false,
-) : BaseEoseManager<T>(client, allKeys) {
-    val sub =
-        requestNewSubscription(
-            object : SubscriptionListener {
-                override fun onEose(
-                    relay: NormalizedRelayUrl,
-                    forFilters: List<Filter>?,
-                ) {
-                    if (invalidateAfterEose) {
-                        invalidateFilters()
-                    }
-                }
-
-                override fun onEvent(
-                    event: Event,
-                    isLive: Boolean,
-                    relay: NormalizedRelayUrl,
-                    forFilters: List<Filter>?,
-                ) {
-                    if (isLive) {
-                        if (invalidateAfterEose) {
-                            invalidateFilters()
-                        }
-                    }
-                }
-            },
-        )
-
-    override fun updateSubscriptions(keys: Set<T>) {
-        val uniqueSubscribedAccounts = keys.distinctBy { distinct(it) }
-        val newFilters = updateFilter(uniqueSubscribedAccounts)?.ifEmpty { null }
-        sub.updateFilters(newFilters?.groupByRelay())
-    }
-
-    abstract fun updateFilter(keys: List<T>): List<RelayBasedFilter>?
-
-    abstract fun distinct(key: T): Any
-}
+typealias SingleSubNoEoseCacheEoseManager<T> = com.vitorpamplona.amethyst.commons.relayClient.eoseManagers.SingleSubNoEoseCacheEoseManager<T>
