+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chats/publicChannels/datasource/subassemblies/FilterMyMessagesToEphemeralChat.kt
@@ -21,32 +21,13 @@
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.publicChannels.datasource.subassemblies
 
 import com.vitorpamplona.amethyst.commons.model.emphChat.EphemeralChatChannel
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.experimental.ephemChat.chat.EphemeralChatEvent
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
 import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterMyMessagesToEphemeralChat as commonsFilterMyMessagesToEphemeralChat
 
 fun filterMyMessagesToEphemeralChat(
     channel: EphemeralChatChannel,
     pubKey: HexKey,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> =
-    channel.relays().toSet().map {
-        RelayBasedFilter(
-            relay = it,
-            filter =
-                Filter(
-                    kinds = listOf(EphemeralChatEvent.KIND),
-                    tags =
-                        if (channel.roomId.id.isBlank()) {
-                            mapOf("d" to listOf("_"))
-                        } else {
-                            mapOf("d" to listOfNotNull(channel.roomId.id))
-                        },
-                    authors = listOf(pubKey),
-                    limit = 50,
-                    since = since?.get(it)?.time,
-                ),
-        )
-    }
+): List<RelayBasedFilter> = commonsFilterMyMessagesToEphemeralChat(channel, pubKey, since)
