+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/communities/datasource/FilterCommunityPosts.kt
@@ -20,79 +20,33 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.communities.datasource
 
-import com.vitorpamplona.quartz.experimental.attestations.attestation.AttestationEvent
 import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
-import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip72ModCommunities.approval.CommunityPostApprovalEvent
 import com.vitorpamplona.quartz.nip72ModCommunities.definition.CommunityDefinitionEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
+import com.vitorpamplona.amethyst.commons.relayClient.filters.CommunityPostKinds as CommonsCommunityPostKinds
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterCommunityPosts as commonsFilterCommunityPosts
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterCommunityPostsFromEverybody as commonsFilterCommunityPostsFromEverybody
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterCommunityPostsFromModerators as commonsFilterCommunityPostsFromModerators
 
-val CommunityPostKinds =
-    listOf(
-        TextNoteEvent.KIND,
-        CommentEvent.KIND,
-        RepostEvent.KIND,
-        GenericRepostEvent.KIND,
-        ClassifiedsEvent.KIND,
-        LongTextNoteEvent.KIND,
-        CommunityPostApprovalEvent.KIND,
-        AttestationEvent.KIND,
-    )
+// Re-exports from commons
+val CommunityPostKinds = CommonsCommunityPostKinds
 
 fun filterCommunityPosts(
     relay: NormalizedRelayUrl,
     community: CommunityDefinitionEvent,
     since: Long?,
-): RelayBasedFilter =
-    RelayBasedFilter(
-        relay = relay,
-        filter =
-            Filter(
-                authors = community.moderatorKeys(),
-                tags = mapOf("a" to listOf(community.addressTag())),
-                kinds = CommunityPostKinds,
-                limit = 500,
-                since = since,
-            ),
-    )
+): RelayBasedFilter = commonsFilterCommunityPosts(relay, community, since)
 
 fun filterCommunityPostsFromModerators(
     relay: NormalizedRelayUrl,
     authors: Set<HexKey>,
     community: CommunityDefinitionEvent,
     since: Long?,
-): RelayBasedFilter =
-    RelayBasedFilter(
-        relay = relay,
-        filter =
-            Filter(
-                authors = authors.sorted(),
-                tags = mapOf("a" to listOf(community.addressTag())),
-                kinds = CommunityPostKinds,
-                limit = 100,
-                since = since,
-            ),
-    )
+): RelayBasedFilter = commonsFilterCommunityPostsFromModerators(relay, authors, community, since)
 
 fun filterCommunityPostsFromEverybody(
     relay: NormalizedRelayUrl,
     community: CommunityDefinitionEvent,
     since: Long?,
-): RelayBasedFilter =
-    RelayBasedFilter(
-        relay = relay,
-        filter =
-            Filter(
-                tags = mapOf("a" to listOf(community.addressTag())),
-                kinds = CommunityPostKinds,
-                limit = 100,
-                since = since,
-            ),
-    )
+): RelayBasedFilter = commonsFilterCommunityPostsFromEverybody(relay, community, since)
