+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/geohash/datasource/FilterPostsByGeohash.kt
@@ -20,68 +20,17 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.geohash.datasource
 
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.home.datasource.nip22Comments.CommentKinds
-import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
-import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip73ExternalIds.location.GeohashId
-import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
-import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
+import com.vitorpamplona.amethyst.commons.relayClient.filters.PostsByGeohashKinds as CommonsPostsByGeohashKinds
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterPostsByGeohash as commonsFilterPostsByGeohash
 
-val PostsByGeohashKinds =
-    listOf(
-        TextNoteEvent.KIND,
-        ChannelMessageEvent.KIND,
-        LongTextNoteEvent.KIND,
-        PollEvent.KIND,
-        ZapPollEvent.KIND,
-        ClassifiedsEvent.KIND,
-        HighlightEvent.KIND,
-        AudioTrackEvent.KIND,
-        AudioHeaderEvent.KIND,
-        WikiNoteEvent.KIND,
-    )
+// Re-exports from commons
+val PostsByGeohashKinds = CommonsPostsByGeohashKinds
 
 fun filterPostsByGeohash(
     geohash: String,
     relays: Set<NormalizedRelayUrl>,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> {
-    val geohashesToFollowMap = mapOf("g" to listOf(geohash))
-    val geohashesScoreMap = mapOf("I" to listOf(GeohashId.toScope(geohash)))
-
-    return relays.flatMap { relay ->
-        val since = since?.get(relay)?.time
-        listOf(
-            RelayBasedFilter(
-                relay = relay,
-                filter =
-                    Filter(
-                        tags = geohashesToFollowMap,
-                        kinds = PostsByGeohashKinds,
-                        limit = 100,
-                        since = since,
-                    ),
-            ),
-            RelayBasedFilter(
-                relay = relay,
-                filter =
-                    Filter(
-                        tags = geohashesScoreMap,
-                        kinds = CommentKinds,
-                        limit = 100,
-                        since = since,
-                    ),
-            ),
-        )
-    }
-}
+): List<RelayBasedFilter> = commonsFilterPostsByGeohash(geohash, relays, since)
