+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/datasource/FilterUserProfileLists.kt
@@ -20,44 +20,15 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.datasource
 
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.experimental.nipA3.PaymentTargetsEvent
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip51Lists.PinListEvent
-import com.vitorpamplona.quartz.nip51Lists.bookmarkList.BookmarkListEvent
-import com.vitorpamplona.quartz.nip51Lists.bookmarkList.OldBookmarkListEvent
-import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
-import com.vitorpamplona.quartz.nip51Lists.hashtagList.HashtagListEvent
-import com.vitorpamplona.quartz.nip51Lists.peopleList.PeopleListEvent
-import com.vitorpamplona.quartz.nip89AppHandlers.recommendation.AppRecommendationEvent
+import com.vitorpamplona.amethyst.commons.relayClient.filters.UserProfileListKinds as CommonsUserProfileListKinds
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterUserProfileLists as commonsFilterUserProfileLists
 
-val UserProfileListKinds =
-    listOf(
-        BookmarkListEvent.KIND,
-        OldBookmarkListEvent.KIND,
-        PinListEvent.KIND,
-        PeopleListEvent.KIND,
-        FollowListEvent.KIND,
-        HashtagListEvent.KIND,
-        AppRecommendationEvent.KIND,
-        PaymentTargetsEvent.KIND,
-    )
+val UserProfileListKinds = CommonsUserProfileListKinds
 
 fun filterUserProfileLists(
     users: Map<NormalizedRelayUrl, Set<String>>,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> =
-    users.map {
-        RelayBasedFilter(
-            relay = it.key,
-            filter =
-                Filter(
-                    kinds = UserProfileListKinds,
-                    authors = it.value.sorted(),
-                    limit = 100,
-                    since = since?.get(it.key)?.time,
-                ),
-        )
-    }
+): List<RelayBasedFilter> = commonsFilterUserProfileLists(users, since)
