+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/datasource/FilterUserProfileMetadata.kt
@@ -20,40 +20,15 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.datasource
 
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.nip01Core.metadata.MetadataEvent
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip02FollowList.ContactListEvent
-import com.vitorpamplona.quartz.nip58Badges.accepted.AcceptedBadgeSetEvent
-import com.vitorpamplona.quartz.nip58Badges.profile.ProfileBadgesEvent
-import com.vitorpamplona.quartz.nip65RelayList.AdvertisedRelayListEvent
+import com.vitorpamplona.amethyst.commons.relayClient.filters.UserProfileMetadataKinds as CommonsUserProfileMetadataKinds
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterUserProfileMetadata as commonsFilterUserProfileMetadata
 
-val UserProfileMetadataKinds =
-    listOf(
-        MetadataEvent.KIND,
-        AdvertisedRelayListEvent.KIND,
-        ContactListEvent.KIND,
-        AcceptedBadgeSetEvent.KIND,
-        ProfileBadgesEvent.KIND,
-    )
+val UserProfileMetadataKinds = CommonsUserProfileMetadataKinds
 
 fun filterUserProfileMetadata(
     users: Map<NormalizedRelayUrl, Set<String>>,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> {
-    if (users.isEmpty()) return emptyList()
-
-    return users.map {
-        RelayBasedFilter(
-            relay = it.key,
-            filter =
-                Filter(
-                    kinds = UserProfileMetadataKinds,
-                    authors = it.value.sorted(),
-                    since = since?.get(it.key)?.time,
-                ),
-        )
-    }
-}
+): List<RelayBasedFilter> = commonsFilterUserProfileMetadata(users, since)
