+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/relay/datasource/FilterPostsByRelay.kt
@@ -20,70 +20,17 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.relay.datasource
 
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
-import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
-import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStorySceneEvent
-import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
-import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
-import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
+import com.vitorpamplona.amethyst.commons.relayClient.filters.PostsByRelayKinds as CommonsPostsByRelayKinds
+import com.vitorpamplona.amethyst.commons.relayClient.filters.PostsByRelayKinds2 as CommonsPostsByRelayKinds2
+import com.vitorpamplona.amethyst.commons.relayClient.filters.filterPostsByRelay as commonsFilterPostsByRelay
 
-val PostsByRelayKinds =
-    listOf(
-        TextNoteEvent.KIND,
-        ChannelMessageEvent.KIND,
-        LongTextNoteEvent.KIND,
-        PollEvent.KIND,
-        LiveActivitiesChatMessageEvent.KIND,
-        ClassifiedsEvent.KIND,
-        HighlightEvent.KIND,
-        WikiNoteEvent.KIND,
-        CommentEvent.KIND,
-    )
-
-val PostsByRelayKinds2 =
-    listOf(
-        InteractiveStorySceneEvent.KIND,
-        AudioTrackEvent.KIND,
-        AudioHeaderEvent.KIND,
-        NipTextEvent.KIND,
-        ZapPollEvent.KIND,
-    )
+val PostsByRelayKinds = CommonsPostsByRelayKinds
+val PostsByRelayKinds2 = CommonsPostsByRelayKinds2
 
 fun filterPostsByRelay(
     relayUrl: NormalizedRelayUrl,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> {
-    val sinceTime = since?.get(relayUrl)?.time
-    return listOf(
-        RelayBasedFilter(
-            relay = relayUrl,
-            filter =
-                Filter(
-                    kinds = PostsByRelayKinds,
-                    limit = 400,
-                    since = sinceTime,
-                ),
-        ),
-        RelayBasedFilter(
-            relay = relayUrl,
-            filter =
-                Filter(
-                    kinds = PostsByRelayKinds2,
-                    limit = 100,
-                    since = sinceTime,
-                ),
-        ),
-    )
-}
+): List<RelayBasedFilter> = commonsFilterPostsByRelay(relayUrl, since)
