+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/relayClient/filters/FilterMyMessagesToPublicChat.kt
@@ -18,14 +18,30 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.datasource
+package com.vitorpamplona.amethyst.commons.relayClient.filters
 
+import com.vitorpamplona.amethyst.commons.model.nip28PublicChats.PublicChatChannel
 import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
+import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.amethyst.commons.relayClient.filters.filterUserProfileLastSeen as commonsFilterUserProfileLastSeen
+import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
+import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
 
-fun filterUserProfileLastSeen(
-    users: Map<NormalizedRelayUrl, Set<String>>,
+fun filterMyMessagesToPublicChat(
+    channel: PublicChatChannel,
+    pubKey: HexKey,
     since: SincePerRelayMap?,
-): List<RelayBasedFilter> = commonsFilterUserProfileLastSeen(users, since)
+): List<RelayBasedFilter> =
+    channel.relays().toSet().map {
+        RelayBasedFilter(
+            relay = it,
+            filter =
+                Filter(
+                    kinds = listOf(ChannelMessageEvent.KIND),
+                    tags = mapOf("e" to listOfNotNull(channel.idHex)),
+                    authors = listOf(pubKey),
+                    limit = 50,
+                    since = since?.get(it)?.time,
+                ),
+        )
+    }
