+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/relayClient/filters/FilterPostsByRelay.kt
@@ -1,0 +1,89 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.relayClient.filters
+
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
+import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
+import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
+import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStorySceneEvent
+import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
+import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
+import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
+import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
+import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
+import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
+import com.vitorpamplona.quartz.nip22Comments.CommentEvent
+import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
+import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
+import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
+import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
+import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
+import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
+import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
+
+val PostsByRelayKinds =
+    listOf(
+        TextNoteEvent.KIND,
+        ChannelMessageEvent.KIND,
+        LongTextNoteEvent.KIND,
+        PollEvent.KIND,
+        LiveActivitiesChatMessageEvent.KIND,
+        ClassifiedsEvent.KIND,
+        HighlightEvent.KIND,
+        WikiNoteEvent.KIND,
+        CommentEvent.KIND,
+    )
+
+val PostsByRelayKinds2 =
+    listOf(
+        InteractiveStorySceneEvent.KIND,
+        AudioTrackEvent.KIND,
+        AudioHeaderEvent.KIND,
+        NipTextEvent.KIND,
+        ZapPollEvent.KIND,
+    )
+
+fun filterPostsByRelay(
+    relayUrl: NormalizedRelayUrl,
+    since: SincePerRelayMap?,
+): List<RelayBasedFilter> {
+    val sinceTime = since?.get(relayUrl)?.time
+    return listOf(
+        RelayBasedFilter(
+            relay = relayUrl,
+            filter =
+                Filter(
+                    kinds = PostsByRelayKinds,
+                    limit = 400,
+                    since = sinceTime,
+                ),
+        ),
+        RelayBasedFilter(
+            relay = relayUrl,
+            filter =
+                Filter(
+                    kinds = PostsByRelayKinds2,
+                    limit = 100,
+                    since = sinceTime,
+                ),
+        ),
+    )
+}
